requirements.txt
streamlit
numpy
Pillow
matplotlib
scipy
pycryptodome